backend/requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-object-actions==5.0.0
django-prometheus==2.3.1
django-timezone-field==7.1
djangorestframework==3.15.2
firebase-admin==6.6.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
kombu==5.5.3
magic-filter==1.0.12
multidict==6.4.3
nest-asyncio==1.6.0
pillow==11.2.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.8.2
pydantic_core==2.20.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==8.1
redis==5.2.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.19.0
gunicorn==23.0.0
PyJWT==2.9.0
